requirements.txt
langchain==0.1.0
langchain_core==0.1.13
langchain_openai==0.0.2.post1
langchainhub==0.1.14
streamlit==1.31.0
python-dotenv==1.0.0
langchain_experimental==0.0.87
numexpr==2.9.0